Check for resizing offset buffers independently
The resize algorithm to increase the capacity of buffers was
tied to checking the remaining capacity of the data buffer. For variable
length fields, the offset buffer needs to be checked independently.

We encountered an array in which a string field which has a
single character. Since this is 1 byte in length, but offsets are 4
bytes in length, we ran out of offset room in the buffer before we had
enough data to force a resize.