Try to pre-import tiledb.cloud before tests
Due to import internals, the first call to `import tiledb.cloud` can have a significant
variance in time. Locally I saw 150-500 ms for a localized call to `tiledb.open` in
`test_bytes_df`. This could explain sc-8256 if `test_bytes` sorts early and is the first
call to `tiledb.open`. Looking at the failing CI CI traces, `test_bytes_*` is run early,
before test_libtiledb, but after `test_fixes` and `test_examples` -- both of which
calls to `tiledb.open`. So this may be a red-herring, but it is the best lead I've
found so far.