Fix SC-13800: properly initialize query in order to estimate results
In return_incomplete mode, prior to this PR the multi_index and df indexer
estimated_result_sizes function would error out until iteration was started,
because the underlying query object was not initialized. In this PR we
fix this issue by initializing the query and updating iteration logic
using a state enum.