Update deadline to 2000, x-ref
After adding some debug info, I'm about 98% sure this is
  `https://github.com/HypothesisWorks/hypothesis/issues/2108` or
  `https://github.com/HypothesisWorks/hypothesis/issues/3369` or similar.

Timing the test itself: the function body in the failing instance completely very quickly. So, barring something extremely spooky, the only place the time could be hiding is in the hypothesis setup for the function (draws are supposed to be excluded but the issues above demonstrate there still appear to be corner cases).