Remove conflict between tiledb.version module and member.
Because the generated `version.py` module would have a fully-qualified
name of `tiledb.version`, but we also used `tiledb.version` as the name
for the `VersionHelper` which contained version information, this meant
that the module `tiledb.version` and the name `tiledb.version` referred
to different things.

This change moves things around slightly to resolve that conflict.