Use object equality check in buffer conversion; fixes sc-33742
The test here fails when run with dask distributed. The array is passed
through dask serialization, and the buffer conversion fails because
the dtype object in the npbuffer.cc line modified here returns False.
This is because the array's dtype is not === to the dtype("O") object
in the parent process after serialization.