Warn when a process that has imported tiledb calls fork() (#1876)
TileDB uses multiple threads and it's easy to run into deadlocks on Linux where `fork` is the default for `multiprocessing` until Python 3.14. We want to see a warning like Python 3.12 emits when fork() is called with multiple threads detected (see the implementation in posixmodule.c). I see three different ways to do this.

1. Do nothing other than document the situation and wait for users to migrate to Python 3.12.
2. Warn about fork() immediately on import of tiledb.
3. Warn when `os.fork()` is called and tiledb has been imported.  

The pros for number one: only needing to reference the change coming in 3.12. The cons: this project currently supports 4 versions of Python before 3.12 and it'll be years before new users are entirely on 3.12. Also, to a first approximation, nobody reads documentation. Perhaps better said as: we often try to read as little documentation as we can get away with.

The pros for number two: super easy to implement and all tiledb users see it. The cons: users would see the warning even if they had no intention of using `fork()`. It's noisy and would have to be actively silenced.

The pros for number three: all tiledb users who call `os.fork()`, intentionally or not, get the warning and no one else. It's not noisy. The cons: it requires monkeypatching`os.fork()` or, suboptimally, using `os.register_at_fork()`.

The only problem with `os.register_at_fork()` is that exceptions and warnings can't be raised from registered callables. We're limited to printing to stderr/stdout. This is a less good experience for developers and the project also currently checks in tests that tiledb doesn't print such output.

Monkeypatching methods of the standard library isn't a perfect solution. We'd all prefer that modules we import incidentally didn't change things in the standard library, right? And there's no guarantee that a module imported after tiledb won't patch `os.fork()` again.

On the other hand, users import tiledb deliberately, to use it. And it's mostly incompatible with `fork()`. The monkeypatch I'm suggesting for `os.fork()` doesn't change behavior of the stdlib function, it only wraps it in a warning. It's as innocuous as such a patch can be.

--- Original commit comments ---

* Add a failing test

We want to get a DeprecationWarning like Python 3.12 emits when
fork() is called with multiple threads detected.

* Monkeypatch os.fork and raise a warning

This is equivalent to the warnings added in Python 3.12. We don't
try to count threads because TileDB is multithreaded.

* Isolate built-in os.fork from in-test patching

* Don't monkeypatch on Windows

* Add a module level flag indicating whether os.fork needs a patch

Drop use of monkeypatch fixture for os.fork isolation. Add some
docs and comments to the two new fixtures and rename the test
module.

* Never access os.fork on windows

* Make isolate_os_fork fixture yield on all platforms

* Add missing win32 skip marks for two tests

* Add another comment about patching and a history entry